Little cleanup around --local

+++ setup.sh
@@ -141,19 +141,19 @@ function _preparation() {
 	apt-get -y install whiptail git sudo curl wget lsof fail2ban apache2-utils vnstat tcl tcl-dev build-essential dirmngr apt-transport-https bc uuid-runtime jq net-tools fortune >> ${log} 2>&1
 	nofile=$(grep "DefaultLimitNOFILE=500000" /etc/systemd/system.conf)
 	if [[ ! "$nofile" ]]; then echo "DefaultLimitNOFILE=500000" >> /etc/systemd/system.conf; fi
-	echo "Cloning swizzin repo to localhost"
   if [[ $local != "true" ]]; then 
+    echo "Cloning swizzin repo to localhost"
 		git clone https://github.com/liaralabs/swizzin.git /etc/swizzin >> ${log} 2>&1
 		#shellcheck source=sources/functions/color_echo
 		. /etc/swizzin/sources/functions/color_echo
 	else
-		RelativeScriptPath=$(dirname "$0")
-		ln -sr "$RelativeScriptPath" /etc/swizzin
-		#shellcheck source=sources/functions/color_echo
-		. /etc/swizzin/sources/functions/color_echo
-		echo_info "WELCOME TO THE WORLD OF THE SWIZ YOUNG PADAWAN\nInstead of cloning from upstream, the directory where the setup script is located is symlinked to /etc/swizzin"
-		echo_info "That directory is relative to your pwd  = $(pwd)/$RelativeScriptPath"
-		echo_warn "Best of luck and please follow the contribution guidelines cheerio"
+    echo "Symlinking $RelativeScriptPath to /etc/swizzin"
+    ln -sr "$RelativeScriptPath" /etc/swizzin
+    echo "We appreciate any and all PRs. please read the Contributing.md as it has a lot of useful tips when working with swizzin."
+    if [[ ! -e /etc/swizzin ]]; then
+      echo "so something fucked up, please join discord and tell us what you did coz that shouldn't have happened"
+      exit 1
+    fi
 	fi
 	ln -s /etc/swizzin/scripts/ /usr/local/bin/swizzin
 	chmod -R 700 /etc/swizzin/scripts
